refactor: ♻️ adds AlertHeaderContent component
